refactor: move withdrawal modal to proper structure
