+++ src/common/components/shared/index.ts
@@ -1,3 +1,3 @@
 export * from './cs-link';
 export * from './popover';
-export * from './modals';
+export * from './modal';
