+++ src/common/components/shared/modal/index.ts
@@ -1,0 +1,1 @@
+export * from './modal-shell';
