--- src/common/components/shared/modals/index.ts
@@ -1,4 +1,0 @@
-export * from './withdrawal';
-export * from './withdrawal-success';
-export * from './withdrawal-error';
-
