+++ src/features/auth/viewmodels/index.ts
@@ -1,3 +1,4 @@
 export * from './stores';
 export * from './use-auth';
+export * from './use-withdrawal-flow';
 export * from './queries';
