+++ src/features/auth/views/components/withdrawal/index.ts
@@ -1,0 +1,3 @@
+export { WithdrawalSuccessModal } from './withdrawal-success-modal';
+export { WithdrawalErrorModal } from './withdrawal-error-modal';
+export { WithdrawalModal } from './withdrawal-modal';
