+++ src/features/auth/views/components/withdrawal/withdrawal-success-modal.tsx
@@ -1,9 +1,8 @@
-import { Button } from '@/common/components/ui/button';
 import { useTranslation } from 'react-i18next';
 
 import CheckIcon from '@/assets/icons/check.svg?react';
-
-import { ModalShell } from './modal-shell';
+import { Button } from '@/common/components/ui/button';
+import { ModalShell } from '@/common/components/shared/modal/modal-shell';
 
 interface WithdrawalSuccessModalProps {
   isOpen: boolean;
